fix: temp fix by only overwriting if not existing
* not true fix, the rsync itself breaks codesigning every time - should install the
pkg directly instead

+++ modules/darwin/kmonad-and-kanata.nix
@@ -38,8 +38,11 @@ in {
     # kmonad setup
     # follows https://github.com/mtoohey31/nixexprs/blob/main/nix-darwin/modules/mtoohey/kmonad.nix
     system.activationScripts.applications.text = pkgs.lib.mkForce (''
-      echo copying dext...
-      ${pkgs.rsync}/bin/rsync -a --delete ${pkgs.Karabiner-DriverKit-VirtualHIDDevice}/Applications/.Karabiner-VirtualHIDDevice-Manager.app /Applications
+      DEST_PATH=${pkgs.Karabiner-DriverKit-VirtualHIDDevice}/Applications/.Karabiner-VirtualHIDDevice-Manager.app
+      if [ ! -d "$DEST_PATH" ]; then
+        echo copying dext...
+        ${pkgs.rsync}/bin/rsync -a --delete ${pkgs.Karabiner-DriverKit-VirtualHIDDevice}/Applications/.Karabiner-VirtualHIDDevice-Manager.app /Applications
+      fi
       echo copying shim...
       cp --no-preserve mode ${pkgs.karabiner-daemon-shim}/bin/karabiner-daemon-shim /Applications/.Karabiner-VirtualHIDDevice-Manager.app/karabiner-daemon-shim
       # make service shim usable
